feat: add order service and improve price handling

+++ Part1_ProceduralToOOP/src/OrderLine.cs
@@ -18,7 +18,7 @@ public class OrderLine
         Quantity = quantity;
     }
 
-    public double CalculateTotal()
+    public decimal CalculateTotal()
     {
         return Product.Price * Quantity;
     }
